Deleted some comments, cleaned up the code a bit
