--- code/src/Time.css
@@ -1,5 +1,0 @@
-.time{
-  color: #CACACA;
-  font-size: 12px;
-  font-weight: bold;
-}
