--- code/src/Time.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import './Time.css'
-
-export const Time = (props) => {
-
-  return (
-    <div className="time">
-      {props.time}
-    </div >
-  )
-
-}
