+++ code/src/components/PostedText.css
@@ -1,4 +1,3 @@
 .posted-text{
-  margin-top:5px;
   font-size: 14px;
 }
